+++ src/main/java/com/minecraftabnormals/upgrade_aquatic/common/entities/GlowSquidEntity.java
@@ -50,7 +50,7 @@ public class GlowSquidEntity extends SquidEntity implements IBucketableEntity {
 		for (int i = 0; i < 30; ++i) {
 			Vector3d vector3d1 = this.rotateVector(new Vector3d((double) this.random.nextFloat() * 0.6D - 0.3D, -1.0D, (double) this.random.nextFloat() * 0.6D - 0.3D));
 			Vector3d vector3d2 = vector3d1.scale(0.3D + (double) (this.random.nextFloat() * 2.0F));
-			((ServerWorld) this.level).sendParticles(UAParticles.GLOW_SQUID_INK.get(), vector3d.x, vector3d.y + 0.5D, vector3d.z, 0, vector3d2.x, vector3d2.y, vector3d2.z, (double) 0.1F);
+			((ServerWorld) this.level).sendParticles(UAParticles.GLOW_SQUID_INK.get(), vector3d.x, vector3d.y + 0.5D, vector3d.z, 0, vector3d2.x, vector3d2.y, vector3d2.z, 0.1F);
 		}
 	}
